requirements.txt

streamlit==1.34.0
torch>=2.6.0
transformers>=4.41.2
Pillow==10.0.0
diffusers==0.19.3
requests>=2.31.0
